final database setup: two objects of database created in MainActivity and SensorlistnerService and both should have the same table name to refer to same database

+++ app/src/main/java/com/example/kushal/mc535/Patient.java
@@ -4,15 +4,16 @@ package com.example.kushal.mc535;
  * Created by RachelDedinsky on 2/23/18.
  */
 
+//Patient objects are stored in the database.
 public class Patient {
     //fields
-    private int patienttimestamp;
+    private long patienttimestamp;
     private int patientxvalues;
     private int patientyvalues;
     private int patientzvalues;
     // constructors
     public Patient() {}
-    public Patient(int timestamp, int xvalues, int yvalues, int zvalues) {
+    public Patient(long timestamp, int xvalues, int yvalues, int zvalues) {
         this.patienttimestamp = timestamp;
         this.patientxvalues = xvalues;
         this.patientyvalues = yvalues;
@@ -20,10 +21,10 @@ public class Patient {
 
     }
     // properties
-    public void setTimestamp(int timeStamp) {
+    public void setTimestamp(long timeStamp) {
         this.patienttimestamp = timeStamp;
     }
-    public int getTimestamp() {
+    public long getTimestamp() {
         return this.patienttimestamp;
     }
     public void setXValues(int x) {
